Add files via upload
This repository contains a Jupyter Notebook titled "DataScienceEcosystem.ipynb" that summarizes the data science ecosystem and presents some of the key tools used in this field. The notebook provides an introduction to popular languages, common libraries, and basic arithmetic operations used in data science. It also includes practical examples of arithmetic expressions and a conversion from minutes to hours. The main objective is to provide an overview of the data science ecosystem and serve as an introductory resource for those looking to delve into this exciting field.